+++ py_src/snex/__init__.py
@@ -1,7 +1,8 @@
 from .etf import set_custom_encoder
-from .interface import BeamModuleProxy, Elixir, call, cast, send
+from .interface import BeamModuleProxy, Elixir, call, cast, io_loop_for_connection, send
 from .logger import LoggingHandler
 from .models import Atom, DistinctAtom, EncodingOpt, EncodingOpts, Term
+from .runner import serve, serve_forever
 
 __all__ = [
     "Atom",
@@ -14,6 +15,9 @@ __all__ = [
     "Term",
     "call",
     "cast",
+    "io_loop_for_connection",
     "send",
+    "serve",
+    "serve_forever",
     "set_custom_encoder",
 ]