+++ py_src/snex/__main__.py
@@ -5,12 +5,8 @@ import asyncio
 import io
 import sys
 from contextlib import suppress
-from typing import TYPE_CHECKING
 
 from . import runner
-
-if TYPE_CHECKING:
-    from .transport import FileLike
 
 sys.stdout = io.TextIOWrapper(sys.stdout.buffer, line_buffering=True, newline="\r\n")
 sys.stderr = io.TextIOWrapper(sys.stderr.buffer, line_buffering=True, newline="\r\n")
@@ -20,9 +16,20 @@ parser.add_argument("--buffer-limit", type=int, default=8 * 1024 * 1024)
 args = parser.parse_args()
 
 
-async def run_loop(erl_in: FileLike, erl_out: FileLike, buffer_limit: int) -> None:
-    reader, writer = await runner.init(erl_in, erl_out, buffer_limit)
-    await runner.run_loop(reader, writer)
+async def serve_forever(
+    erl_in: io.FileIO,
+    erl_out: io.FileIO,
+    buffer_limit: int,
+) -> None:
+    loop = asyncio.get_running_loop()
+
+    reader = asyncio.StreamReader(limit=buffer_limit, loop=loop)
+    protocol = asyncio.StreamReaderProtocol(reader, loop=loop)
+    await loop.connect_read_pipe(lambda: protocol, erl_in)
+    transport, _ = await loop.connect_write_pipe(asyncio.Protocol, erl_out)
+    writer = asyncio.StreamWriter(transport, protocol, reader, loop=loop)
+
+    await runner.serve_forever(reader, writer)
 
 
 with (
@@ -30,4 +37,4 @@ with (
     open(4, "wb", 0) as erl_out,
     suppress(asyncio.CancelledError),
 ):
-    asyncio.run(run_loop(erl_in, erl_out, args.buffer_limit))
+    asyncio.run(serve_forever(erl_in, erl_out, args.buffer_limit))